pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mep"
version = "0.2.0"
description = "Muon Equilibrium Propagation: Biologically plausible deep learning"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["deep-learning", "equilibrium-propagation", "biologically-plausible", "energy-based-models"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]
research = [
    "wandb>=0.15.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "pandas>=1.5.0",
    "scipy>=1.9.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
]
all = ["mep[dev,research]"]

[project.scripts]
mep-benchmark = "mep.cli:benchmark"

[tool.setuptools.packages.find]
include = ["mep*"]
exclude = ["data", "tests*", "benchmarks*", "experiments*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=mep --cov-report=html --cov-report=term"
